Add files via upload
Python Task List is a python file which is containing of 2 python programs. The first one is to find the missing number in array. Here we have to find the missing number in array by using function Missing_Number.
The second one is of the Valid Parentheses Checker. Here we have to write the function to check whether the given string containing only () , {} , [] is balanced.